sv3-2022-jvjbf-harmadik-reszvizsga/src/main/java/guinness/Recorder.java
package guinness;

import java.util.Objects;

public class Recorder {

	private int id;
	private String name;

	public Recorder(int id, String name) {
		this.id = id;
		this.name = name;
	}

	public int getId() {
		return id;
	}

	public String getName() {
		return name;
	}

	@Override
	public boolean equals(Object o) {
		if (this == o) return true;
		if (o == null || getClass() != o.getClass()) return false;
		Recorder recorder = (Recorder) o;
		return id == recorder.id;
	}

	@Override
	public int hashCode() {
		return Objects.hash(id);
	}
}
